+++ examples/websockets/falcon.rb
@@ -6,7 +6,7 @@
 
 require "falcon/environment/self_signed_tls"
 require "falcon/environment/rack"
-require "falcon/environment/supervisor"
+require "async/service/supervisor"
 
 service "websockets.localhost" do
 	include Falcon::Environment::SelfSignedTLS
@@ -14,5 +14,5 @@ service "websockets.localhost" do
 end
 
 service "supervisor" do
-	include Falcon::Environment::Supervisor
+	include Async::Service::Supervisor::Environment
 end